API uploads images via URL
This enhancement give the images API the ability to create images from URLs.  A
simple check determines whether or not the attachment parameter is a URL.  If it
is a URL, OpenURI is used to download the image and it subsequently undergoes
the standard post upload processing (via paperclip).  If it is not a URL, then
the attachment parameter is handled in the same manner it was prior to this
commit.